organized .py files into appropiate folders
