+++ .gitignore
@@ -118,7 +118,7 @@ var
 sdist
 develop-eggs
 .installed.cfg
-lib
+
 lib64
 __pycache__
