+++ DeepSleep/modules/__init__.py
@@ -1,0 +1,3 @@
+#from .AnaDict import AnaDict
+#from .getdata import getData
+#from .processAna import processAna
