+++ DeepSleep/submitjobs.py
@@ -3,7 +3,7 @@
 #
 import sys
 import os
-import deepsleepcfg as cfg
+import cfg.deepsleepcfg as cfg
 #
 # add parsargs at some point for year, rundata, minibatch
 # submit a job for each background / signal / data sample
@@ -11,7 +11,7 @@ samples = cfg.ZHbbFitCfg[1]
 log_dir = 'log/'
 os.system('rm log/*')
 for sample in samples:
-    command = f'qsub -o {log_dir}{sample}.out -e {log_dir}{sample}.err -v sample={sample} runAna.sh'
+    command = f'qsub -l nodes=1:ppn=1 -o {log_dir}{sample}.out -e {log_dir}{sample}.err -v sample={sample} runAna.sh'
     print(command)
     os.system(command)
